Updated consolidated code.

+++ file_reader.py
@@ -1,33 +1,67 @@
-import os
+import os.path, time
 
 # Sort jobs applied and count amount of jobs
 def sort_jobs(inputFile):
     appliedCount = 0
     rejections = 0
 
+    # Current Date
+    currentDate = time.strftime("%m/%d/%Y")
+
     # Put text file contents into a list to sort
     with open(inputFile, 'r') as inFile:
-        firstLine = inFile.readline()
-        contentList = [companyNames.strip('\n') for companyNames in inFile.readlines()] 
+        if currentDate in inFile.read():
+            for line in inFile:
+                if currentDate in line:
+                    existingFile.write("\n{}".format(currentDate));
+                    newJobsList = [line.strip('\n') for line in existingFile]
+        else:
+            firstLine = inFile.readline()
+            contentList = [companyNames.strip('\n') for companyNames in inFile.readlines()]
 
     # Sort List alphabetically
     contentList.sort()
+    newJobsList.sort()
 
-    with open("D:\Software\Google Drive\Documents\jobs_applied_sorted.txt", 'w') as outFile:
-        outFile.write(firstLine + '\n')
-        for names in contentList:
-            if('-X' in names):
-                rejections += 1
-            if '/' not in names:
-                if (names != ''):
-                    outFile.write(names + '\n')
-                    appliedCount += 1
+    if not os.path.exists("D:\Software\Google Drive\Documents\jobs_applied_sorted.txt"):
+
+        with open("D:\Software\Google Drive\Documents\jobs_applied_sorted.txt", 'w') as outFile:
+            outFile.write(firstLine + '\n')
+            update_count_rejections_and_applied(contentList, rejections, appliedCount, outFile)
+            
+    elif os.path.exists("D:\Software\Google Drive\Documents\jobs_applied_sorted.txt"):
+
+        if currentDate not in open("D:\Software\Google Drive\Documents\jobs_applied_sorted.txt", 'r').read():
+
+            # Means that text file is updated and should just update the existing file with the new info
+            with open("D:\Software\Google Drive\Documents\jobs_applied_sorted.txt", 'a') as existingFile:
+                update_coint_rejections_and_applied(newJobsList, rejections, appliedCount, outFile)
                 
-        outFile.write("\nApplied: {}\n".format(appliedCount))
-        outFile.write("Rejections: {}".format(rejections))
-
+          #  for names in newJobsList:
+        #     if('-X' in names):
+         #           rejections += 1
+          #      if '/' not in names:
+           #         if (names != ''):
+            #            existingFile.write(names + '\n')
+             #           appliedCount += 1
+                    
+            #existingFile.write("\nApplied: {}\n".format(appliedCount))
+            #existingFile.write("Rejections: {}".format(rejections))
+                    
     inFile.close()
     outFile.close()
+
+def update_count_rejections_and_applied(list, rejects, applied, outFile):
+    for names in list:
+        if('-X' in names):
+            rejects += 1
+        if '/' not in names:
+            if (names != ''):
+                outFile.write(names + '\n')
+                applied += 1
+            
+    outFile.write("\nApplied: {}\n".format(applied))
+    outFile.write("Rejections: {}".format(rejects))
     
 def print_list(inputList):
     for x in inputList:
@@ -35,7 +69,7 @@ def print_list(inputList):
         
 def main():
     providedFilename = input("Enter fileName: ")
-    sort_jobs(providedFileName + ".txt")
+    sort_jobs(providedFilename + ".txt")
     
 
 if __name__ == "__main__":
